# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Oracle Factory server.",
        "title": "Oracle Factory API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/smartContractListeners": {
            "get": {
                "description": "Retrieves all smart contract listener of the user signed in. This endpoint will be called from the frontend, to display information about all smart contract listeners of the user signed in.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smartContractListener"
                ],
                "summary": "Retrieves all smart contract listener of the user signed in.",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a smart contract listeners for a user. This service will be called by the frontend to when a user filled out the smart contract listener form.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smartContractListener"
                ],
                "summary": "Creates a smart contract listeners for a user",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/smartContractListeners/{smartContractListenerID}": {
            "get": {
                "description": "Retrieves a smart contract listener. This endpoint will be called from the frontend, to display information about a smart contract listener.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smartContractListener"
                ],
                "summary": "Retrieves a smart contract listener",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the smart contract listener to send data to.",
                        "name": "smartContractListenerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/smartContractListeners/{smartContractListenerID}/events": {
            "post": {
                "description": "Handles the event send from a smart contract provider. This endpoint will be called from the BLF, that provides data to the artifact.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "smartContractListener"
                ],
                "summary": "Handles the event send from a smart contract provider",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "the ID of the smart contract listener to send data to.",
                        "name": "smartContractListenerID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/user": {
            "get": {
                "description": "Retrieve the signed in user. This will be called by the frontend to get all information about the user signed in.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Retrieve signed in user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a user. This will be called from the frontend to update the settings",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Update User",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
        "responses.TokenResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    }
}